Typst function

Generates Typst code with element. Named arguments are included as Typst keyword arguments (e.g. level, numbering), and unnamed arguments are used as the heading content. But please note that this behavior might be different in some functions to respect true Typst syntax (such as #table() for instance).

Usage

assert(...)

bytes(...)

datetime(...)

duration(...)

eval_(...)

panic(...)

regex(...)

repr(...)

str_(...)

symbol(...)

decimal(...)

type(...)

version(...)

place(...)

align(...)

h(...)

v(...)

block(...)

box(...)

colbreak(...)

stack(...)

grid(...)

hide(...)

pagebreak(...)

move(...)

pad(...)

page(...)

scale(...)

skew(...)

bibliography(...)

cite(...)

list_(...)

document(...)

emph(...)

figure(...)

footnote(...)

heading(...)

enum(...)

outline(...)

par(...)

parbreak(...)

quote_(...)

strong(...)

table_(...)

text(...)

linebreak(...)

highlight(...)

lower(...)

overline(...)

smallcaps(...)

smartquote(...)

strike(...)

sub_(...)

super(...)

underline(...)

upper(...)

raw_(...)

image(...)

circle(...)

ellipse(...)

line(...)

rect(...)

square(...)

Arguments

...

Named or unnamed arguments passed to the Typst function. Named arguments become Typst parameters (e.g. level = 2 → level: 2), while unnamed arguments form the body inside [...].

Value

A character string containing the generated Typst code.

Note

Does not work yet because label type is not implemented yet. See https://github.com/y-sunflower/rtyp/issues/10

Examples

heading(level = 2, numbering = "1.1", "Hello world")
#> [1] "#heading(level: 2, numbering: \"1.1\")[Hello world]"
text(font = "Libertinus Serif", "Hello world")
#> [1] "#text(font: \"Libertinus Serif\")[Hello world]"
text("No arguments")
#> [1] "#text[No arguments]"

assert(1 < 2, message = "math broke")
#> [1] "#assert(message: \"math broke\", TRUE)"
bytes("more cookies...")
#> [1] "#bytes(\"more cookies...\")"
datetime(year = 2012, month = 8, day = 3)
#> [1] "#datetime(year: 2012, month: 8, day: 3)"
duration(days = 3, hours = 12)
#> [1] "#duration(days: 3, hours: 12)"
eval_("*Markup!*", mode = "markup")
#> [1] "#eval(mode: \"markup\", \"*Markup!*\")"
panic("this is wrong")
#> [1] "#panic(\"this is wrong\")"
regex("\\d+")
#> [1] "#regex(\"\\d+\")"
repr("hello")
#> [1] "#repr(\"hello\")"
str_("hello")
#> [1] "#str(\"hello\")"
symbol("10")
#> [1] "#symbol(\"10\")"
decimal("10")
#> [1] "#decimal(\"10\")"
type("10")
#> [1] "#type(\"10\")"
version(1, 2, 3, 4)
#> [1] "#version(1, 2, 3, 4)"
version(c(1, 2, 3, 4))
#> [1] "#version((1, 2, 3, 4))"
place(dx = pt(50), "pasta")
#> [1] "#place(dx: 50pt)[pasta]"
align(center + horizon, "hey")
#> [1] "#align(center + horizon)[hey]"
h(fr(3))
#> [1] "#h(\"3\")"
v(percent(3), weak = TRUE)
#> [1] "#v(weak: true, \"3\")"
block(breakable = FALSE, width = percent(50), height = "auto", fill = aqua, "cookies")
#> [1] "#block(breakable: false, width: 50%, height: auto, fill: aqua)[cookies]"
box(clip = FALSE, width = percent(50), height = "auto", fill = red, "ice cream")
#> [1] "#box(clip: false, width: 50%, height: auto, fill: red)[ice cream]"
colbreak(weak = TRUE)
#> [1] "#colbreak(weak: true)"
stack(dir = ttb, "hey", "you", "!")
#> [1] "#stack(dir: ttb, \"hey\", \"you\", \"!\")"
grid(columns = 2, rows = 2, gutter = pt(3), "hey", "cookies", "ice", "cream")
#> [1] "#grid(columns: 2, rows: 2, gutter: 3pt, \"hey\", \"cookies\", \"ice\", \"cream\")"
hide("this is hidden")
#> [1] "#hide[this is hidden]"
pagebreak("this is hidden")
#> [1] "#pagebreak(\"this is hidden\")"
move(dx = pt(3), dy = pt(5), "pasta")
#> [1] "#move(dx: 3pt, dy: 5pt, \"pasta\")"
pad(x = percent(0) + pt(5), y = percent(10) + pt(0), image("file.svg"))
#> [1] "#pad(x: 0% + 5pt, y: 10% + 0pt)[#image(\"file.svg\")]"
page(
 flipped = TRUE,
 columns = 2,
 fill = red,
 place(
   top + left,
   dx = pt(-5),
   rect(fill = blue, radius = pt(2), "yooooo")
 )
)
#> [1] "#page(flipped: true, columns: 2, fill: red)[#place(top + left, dx: -5pt)[#rect(fill: blue, radius: 2pt)[yooooo]]]"
scale(x = percent(-100), "This is mirrored")
#> [1] "#scale(x: -100%)[This is mirrored]"
skew(ax = deg(-12), "This is some fake italic text.")
#> [1] "#skew(ax: -12deg)[This is some fake italic text.]"
highlight(extent = pt(3), "hey")
#> [1] "#highlight(extent: 3pt)[hey]"
lower("hey")
#> [1] "#lower[hey]"
raw_("print('hello world')", lang = "python", block = TRUE)
#> [1] "#raw(lang: \"python\", block: true, \"print('hello world')\")"